% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaco_data.R
\name{phaco_data}
\alias{phaco_data}
\title{phaco_data : Charger les données qui sont utilisées dans phacochr}
\usage{
phaco_data(data = NULL, path_data = NULL)
}
\arguments{
\item{data}{Nom des données à charger: "rues" (les rues BeST prétaitées pour l'usage de phacochr), "adresses" (les adresses BeST prétaitées pour l'usage de phacochr), "sec" (les secteurs statistiques), "communes_bel" (les communes de Belgiques), "provinces" (les provinces de Belgiques), "regions" (les régions de Belgiques), "belgique" (les limites de la Belgiques), "rbc" (la Région de Bruxelles-Capitale), "communes_bxl" (les communes de la Région de Bruxelles-Capitale), "quartiers_bxl" (les quartiers monitoring de Bruxelles), "rbc" (les limites de la Région de Bruxelles-Capitale), "sec_bxl" (les secteurs statistiques de la Région des Bruxelles-Capitale). Par défaut: NULL}

\item{path_data}{Chemin absolu vers le dossier où se trouve le données. Par défaut data_path = NULL et phacochr trouve le dossier d'installation choisi par défaut.}
}
\description{
Cette fonction permet de charger les données utilisées par phacochr et qui sont stockée dans l'ordinateur.
}
\examples{
\dontrun{
# Rues BeST
rues <- phaco_data("rues")
# Adresses BeST
adresses <- phaco_data("adresses")
# Secteurs Statistiques
sec<-phaco_data("sec")
# Communes Belgique
communes_bel <- phaco_data("communes_bel")
# Provinces
provinces <- phaco_data("provinces")
# Regions
regions <- phaco_data("regions")
# Belgique
belgique <- phaco_data("belgique")
# Région de Bruxelles-Capital
rbc <- phaco_data("rbc")
# Communes de la Région de Bruxelles-Capitale
communes_bxl <- phaco_data("communes_bxl")
#  Quartier monitoring de la Région de Bruxelles-Capitale
quartiers_bxl <- phaco_data("quartiers_bxl")
# Limites de la Région de Bruxelles-Capital
rbc <- phaco_data("rbc")
# Secteurs Statistiques de la Région de Bruxelles-Capitale
sec_bxl <- phaco_data("sec_bxl")
}


}
